Add and fix tests
