+++ frontend/src/app/features/subsidiary/workplace-users/view-subsidiary-workplace-users.component.ts
@@ -2,10 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { JourneyType } from '@core/breadcrumb/breadcrumb.model';
 import { Establishment } from '@core/model/establishment.model';
-import { UserDetails, UserPermissionsType } from '@core/model/userDetails.model';
 import { BreadcrumbService } from '@core/services/breadcrumb.service';
-import { PermissionsService } from '@core/services/permissions/permissions.service';
-import { TabsService } from '@core/services/tabs.service';
 import { UserService } from '@core/services/user.service';
 import { ParentSubsidiaryViewService } from '@shared/services/parent-subsidiary-view.service';
 
@@ -15,18 +12,12 @@ import { ParentSubsidiaryViewService } from '@shared/services/parent-subsidiary-
 })
 export class ViewSubsidiaryWorkplaceUsersComponent implements OnInit {
   public workplace: Establishment;
-  public users: Array<UserDetails> = [];
-  public canAddUser: boolean;
-  public canViewUser: boolean;
-  public userPermissionsTypes: UserPermissionsType[];
-
+  
   constructor(
     private route: ActivatedRoute,
     private userService: UserService,
-    private permissionsService: PermissionsService,
     private breadcrumbService: BreadcrumbService,
     private parentSubsidiaryViewService: ParentSubsidiaryViewService,
-    private tabsService: TabsService,
   ) {}
 
   ngOnInit(): void {
