+++ frontend/src/app/features/workplace/users/users.component.ts
@@ -4,7 +4,6 @@ import { JourneyType } from '@core/breadcrumb/breadcrumb.model';
 import { Establishment } from '@core/model/establishment.model';
 import { UserDetails, UserPermissionsType } from '@core/model/userDetails.model';
 import { BreadcrumbService } from '@core/services/breadcrumb.service';
-import { PermissionsService } from '@core/services/permissions/permissions.service';
 import { UserService } from '@core/services/user.service';
 
 @Component({
@@ -13,7 +12,6 @@ import { UserService } from '@core/services/user.service';
 })
 export class UsersComponent implements OnInit {
   public workplace: Establishment;
-  public showSecondUserBanner: boolean;
   public users: UserDetails[];
   public canAddUser: boolean;
   public canViewUser: boolean;
@@ -21,7 +19,6 @@ export class UsersComponent implements OnInit {
 
   constructor(
     private route: ActivatedRoute,
-    private permissionsService: PermissionsService,
     private breadcrumbService: BreadcrumbService,
     private userService: UserService,
   ) {}
