+++ frontend/src/app/shared/components/user-accounts-summary/user-accounts-summary.component.spec.ts
@@ -10,20 +10,20 @@ import { MockPermissionsService } from '@core/test-utils/MockPermissionsService'
 import { EditUser, ReadUser } from '@core/test-utils/MockUserService';
 import { UserAccountsSummaryComponent } from '@shared/components/user-accounts-summary/user-accounts-summary.component';
 import { SharedModule } from '@shared/shared.module';
-import { render } from '@testing-library/angular';
+import { queryByTestId, render } from '@testing-library/angular';
 
 import { Establishment } from '../../../../mockdata/establishment';
 
 describe('UserAccountsSummaryComponent', () => {
-  const setup = async (showBanner = false, overLimit = false) => {
-    const { fixture, getByText, getAllByText, getByTestId, queryByText } = await render(UserAccountsSummaryComponent, {
+  const setup = async (showBanner = undefined, overLimit = false, isParentUsers = false) => {
+    const { fixture, getByText, getAllByText, getByTestId, queryByText, queryByTestId } = await render(UserAccountsSummaryComponent, {
       imports: [SharedModule, RouterModule, RouterTestingModule, HttpClientTestingModule],
       declarations: [],
-      componentProperties: { workplace: Establishment, showSecondUserBanner: showBanner },
+      componentProperties: { workplace: Establishment, showSecondUserBanner: showBanner, isParentUsers: isParentUsers },
       providers: [
         {
           provide: PermissionsService,
-          useFactory: MockPermissionsService.factory(['canAddUser']),
+          useFactory: MockPermissionsService.factory(['canAddUser', 'canViewUser']),
           deps: [HttpClient, Router, UserService],
         },
         {
@@ -42,12 +42,51 @@ describe('UserAccountsSummaryComponent', () => {
     });
     const component = fixture.componentInstance;
 
-    return { component, fixture, getByText, getAllByText, getByTestId, queryByText };
+    return { component, fixture, getByText, getAllByText, getByTestId, queryByText, queryByTestId };
   };
 
   it('should render a User Account Summary Workplace Component', async () => {
     const { component } = await setup();
     expect(component).toBeTruthy();
+  });
+
+  it('should show the workplace name and id, and the number of users', async () => {
+    const { getByTestId, getByText } = await setup(false, false, true);
+
+    const workplaceInfo = getByTestId('workplace-name');
+    const numberOfUsers = getByText('Users (1)');
+
+    expect(workplaceInfo).toBeTruthy();
+    expect(numberOfUsers).toBeTruthy();
+  });
+
+  it('should not show the workplace name and id, instead it should only show the number of users', async () => {
+    const { queryByText, getByText } = await setup();
+
+    const workplaceInfo = queryByText('workplace-name');
+    const numberOfUsers = getByText('Users (1)');
+
+    expect(workplaceInfo).toBeFalsy();
+    expect(numberOfUsers).toBeTruthy();
+  });
+
+  it('should show Add a user button with the correct href when the canAddUser permission is true', async () => {
+    const { component, getByText } = await setup();
+
+    const addUserButton = getByText('Add a user');
+    const workplaceId = component.workplace.uid;
+
+    expect(addUserButton).toBeTruthy();
+    expect(addUserButton.getAttribute('href')).toEqual(`/workplace/${workplaceId}/user/create`);
+  });
+
+  it('should not show Add a user button when the canAddUser permission is false', async () => {
+    const { component, fixture, queryByText } = await setup();
+
+    component.canAddUser = false;
+    fixture.detectChanges();
+
+    expect(queryByText('Add a user')).toBeFalsy();
   });
 
   it('should still show Add User if existing user doesnt have a name', async () => {
@@ -85,6 +124,61 @@ describe('UserAccountsSummaryComponent', () => {
 
     expect(queryByText(addUserText, { exact: false })).toBeFalsy();
   });
+
+  it('should show add user banner if there is 1 user and canAddUser permission is true', async () => {
+    const { queryByText } = await setup();
+
+    const addUserText =
+      'Adding a second user will give Skills for Care another person to contact at your workplace should you be unavailable.';
+
+    expect(queryByText(addUserText, { exact: false })).toBeTruthy();
+  });
+
+  it('should not show add user banner if there is 1 user and canAddUser permission is false', async () => {
+    const { component, fixture, queryByText } = await setup();
+
+    component.canAddUser = false;
+    component.setShowSecondUserBanner();
+    fixture.detectChanges();
+
+    const addUserText =
+      'Adding a second user will give Skills for Care another person to contact at your workplace should you be unavailable.';
+
+    expect(queryByText(addUserText, { exact: false })).toBeFalsy();
+  });
+
+  it('should not show add user banner if there is more than 1 user', async () => {
+    const { queryByText } = await setup(undefined, true);
+
+    const addUserText =
+      'Adding a second user will give Skills for Care another person to contact at your workplace should you be unavailable.';
+
+    expect(queryByText(addUserText, { exact: false })).toBeFalsy();
+  });
+
+  it('should show the users full name as a link with the correct href when the canViewUser permissions is true', async () => {
+    const { component, getByTestId } = await setup();
+    const usernameLink = getByTestId('username-link');
+
+    const establishmentId = component.workplace.uid;
+    const userId = component.users[0].uid;
+
+    expect(usernameLink).toBeTruthy();
+    expect(usernameLink.getAttribute('href')).toEqual(`/workplace/${establishmentId}/user/${userId}`);
+  });
+
+  it('should show the users fullname not as a link when the canViewUser permission is false', async () => {
+    const { component, fixture, queryByTestId, getByText } = await setup();
+
+    component.canViewUser = false;
+    fixture.detectChanges();
+
+    const fullname = component.users[0].fullname;
+
+    expect(queryByTestId('username-link')).toBeFalsy();
+    expect(getByText(fullname)).toBeTruthy();
+  });
+
 
   describe('Permissions column', () => {
     it('should have permission as Primary edit and WDF when user isPrimary and canManageWdfClaims are true and role is Edit', async () => {
